# Conflicts:
#   backend/.openapi-generator/swaggerDocumentation.yaml-default.sha256
#   backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/config/SecurityConfig.java
#   swaggerDocumentation.yaml

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/interfaces/ArticlesApi.java
@@ -112,7 +112,7 @@ public interface ArticlesApi {
         produces = { "application/json" }
     )
     default ResponseEntity<ArticleDto> articlesIdGet(
-        @Parameter(name = "id", description = "ID of the user that is retreived", required = true, schema = @Schema(description = "")) @PathVariable("id") Integer id
+        @Parameter(name = "id", description = "ID of the user that is retreived", required = true, schema = @Schema(description = "")) @PathVariable("id") Long id
     ) {
         getRequest().ifPresent(request -> {
             for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
